Fixes #21
We now use Stage and Actor in a more reasonable way. This also makes the
input handling flow a lot more logical and easier to maintain.

Agents are now an extension of Actor. These have set values for
positions and dimensions. Agents now also render using sprites rather
then shapes. The sprites are instead 'tinted' with the agents set color.

This means we can create a lot nicer sprites that may then be tinted
during runtime.